Create convert sub-package
