+++ src/main/java/net/imagej/omero/legacy/convert/IJROIToOverlay.java
@@ -1,5 +1,5 @@
 
-package net.imagej.omero.legacy;
+package net.imagej.omero.legacy.convert;
 
 import ij.gui.Overlay;
