+++ src/main/java/net/imagej/omero/legacy/convert/IJROIToROITree.java
@@ -1,5 +1,5 @@
 
-package net.imagej.omero.legacy;
+package net.imagej.omero.legacy.convert;
 
 import java.util.Collections;
 
@@ -8,18 +8,21 @@ import net.imagej.roi.ROITree;
 import net.imglib2.roi.MaskPredicate;
 
 import org.scijava.convert.AbstractConverter;
+import org.scijava.convert.ConvertService;
 import org.scijava.convert.Converter;
+import org.scijava.plugin.Parameter;
 import org.scijava.plugin.Plugin;
 
 /**
- * Converts a {@link MaskPredicate} to a {@link ROITree}.
+ * Converts a {@link ij.gui.Roi} to a {@link ROITree}.
  *
  * @author Alison Walter
  */
 @Plugin(type = Converter.class)
-public class MaskPredicateToROITree extends
-	AbstractConverter<MaskPredicate<?>, ROITree>
-{
+public class IJROIToROITree extends AbstractConverter<ij.gui.Roi, ROITree> {
+
+	@Parameter
+	private ConvertService convert;
 
 	@Override
 	@SuppressWarnings("unchecked")
@@ -30,8 +33,9 @@ public class MaskPredicateToROITree extends
 			throw new IllegalArgumentException("Expected " + getOutputType() +
 				" but received " + dest);
 
+		final MaskPredicate<?> mp = convert.convert(src, MaskPredicate.class);
 		final ROITree tree = new DefaultROITree();
-		tree.addROIs(Collections.singletonList((MaskPredicate<?>) src));
+		tree.addROIs(Collections.singletonList(mp));
 		return (T) tree;
 	}
 
@@ -41,9 +45,8 @@ public class MaskPredicateToROITree extends
 	}
 
 	@Override
-	@SuppressWarnings({ "unchecked", "rawtypes" })
-	public Class<MaskPredicate<?>> getInputType() {
-		return (Class) MaskPredicate.class;
+	public Class<ij.gui.Roi> getInputType() {
+		return ij.gui.Roi.class;
 	}
 
 }
